[BOOK-434] refactor: CircuitGraph AppGraph 에서 분리
